+++ src/lib/sketch/onboard/Sides_Complete.svelte
@@ -1,6 +1,13 @@
 <script lang="ts">
 	import type {Consent_Principle} from '$lib/sketch/onboard/consent';
-	import {point_left, point_right, arrow_right} from '$lib/ui/icons';
+	import {
+		hand_point_left,
+		hand_point_right,
+		arrow_right,
+		face_think,
+		face_blush,
+		face_wink,
+	} from '$lib/ui/icons';
 
 	export let principle: Consent_Principle;
 	export let consentful_on_left_side: boolean;
@@ -15,7 +22,7 @@
 
 <div class="sides-complete">
 	<div class="icon">
-		{#if correct === null}🤔{:else if correct}😊{:else}😉{/if}
+		{#if correct === null}{face_think}{:else if correct}{face_blush}{:else}{face_wink}{/if}
 	</div>
 	<div class="prompt">
 		{#if correct === null}
@@ -28,8 +35,8 @@
 	</div>
 	<div class="buttons">
 		{#if correct === null}
-			<button on:click={() => choose('left')}>{point_left}</button>
-			<button on:click={() => choose('right')}>{point_right}</button>
+			<button on:click={() => choose('left')}>{hand_point_left}</button>
+			<button on:click={() => choose('right')}>{hand_point_right}</button>
 		{:else}
 			<button on:click={() => done()}>continue {arrow_right}</button>
 		{/if}